Add LegacyController & Controller to handle request intra platform
