--- common/oatbox/http/ResponseEmitter.php
@@ -1,59 +1,0 @@
-<?php
-/**
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; under version 2
- * of the License (non-upgradable).
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- *
- * Copyright (c) 2019 (original work) Open Assessment Technologies SA
- *
- */
-
-namespace oat\oatbox\http;
-
-use Psr\Http\Message\ResponseInterface;
-
-/**
- * Class ResponseEmitter
- * @package oat\oatbox\http
- * @author Moyon Camille <camille@taotesting.com>
- */
-class ResponseEmitter
-{
-    /**
-     * Emit a Http Response to client
-     *
-     * @see https://github.com/http-interop/response-sender
-     * @param ResponseInterface $response
-     */
-    public function __invoke(ResponseInterface $response)
-    {
-        $http_line = sprintf('HTTP/%s %s %s',
-            $response->getProtocolVersion(),
-            $response->getStatusCode(),
-            $response->getReasonPhrase()
-        );
-        header($http_line, true, $response->getStatusCode());
-        foreach ($response->getHeaders() as $name => $values) {
-            foreach ($values as $value) {
-                header("$name: $value", false);
-            }
-        }
-        $stream = $response->getBody();
-        if ($stream->isSeekable()) {
-            $stream->rewind();
-        }
-        while (!$stream->eof()) {
-            echo $stream->read(1024 * 8);
-        }
-    }
-}
